reworked modPortal install and modPortal Login (also use json instead of formdata)
